backend/gpsSimulator/requirements.txt
Django==4.2.11
fastapi==0.115.0
uvicorn[standard]==0.30.1

pymobiledevice3==7.8.1
psutil==5.9.8

requests==2.31.0
polyline==2.0.2

python-dotenv==1.0.1
httpx==0.27.0
pycountry==23.12.11

google-auth 
google-auth-oauthlib 
django-cors-headers
mysqlclient